BaytesSMC

Documentation for BaytesSMC.

BaytesSMC.BaytesSMC

BaytesSMC.CorrelationBuffer

BaytesSMC.SMC

BaytesSMC.SMC2Kernel

BaytesSMC.SMCBuffer

BaytesSMC.SMCConstructor

BaytesSMC.SMCDefault

BaytesSMC.SMCDiagnostics

BaytesSMC.SMCParticles

BaytesSMC.SMCParticles

BaytesSMC.SMCParticles

BaytesSMC.SMCTune

BaytesCore.infer

BaytesCore.jitter!

BaytesCore.propagate!

BaytesCore.propagate!

BaytesCore.propose!

BaytesCore.resample!

BaytesCore.results

BaytesCore.weight!

BaytesSMC.SMCreweight

BaytesSMC.SMCweight

BaytesSMC.compute_batchsize

BaytesSMC.compute_ρ!

BaytesSMC.infer_generated

BaytesSMC.jitterkernel

BaytesSMC.predict!

BaytesSMC.reweight!

BaytesSMC.BaytesSMC — Module

Sequential Monte Carlo module - MCMC and PF algorithm compositionally used

BaytesSMC.CorrelationBuffer — Type

struct CorrelationBuffer{T<:AbstractFloat}

Stores buffers for computing correlation of model parameter particles.

Fields

ρ::Vector{T} where T<:AbstractFloat

Correlation between old and rejuvenated parameter

pairs::Array{Vector{T}, 1} where T<:AbstractFloat

Pairs of old and new parameter samples - for the same parameter

BaytesSMC.SMC — Type

struct SMC{A<:SMCParticles, B<:SMCTune} <: BaytesCore.AbstractAlgorithm

SMC Algorithm.

Fields

particles::SMCParticles

tune::SMCTune

BaytesSMC.SMC2Kernel — Type

struct SMC2Kernel{P<:BaytesFilters.ParticleFilter, M<:BaytesPMCMC.PMCMC} <: BaytesCore.AbstractAlgorithm

SMC2 Kernel, consisting of a kernel for the particle trajectory and kernel for all model parameter.

Fields

pf::BaytesFilters.ParticleFilter

Kernel for (online) state propagation over time.

pmcmc::BaytesPMCMC.PMCMC

Kernel for jittering step.

BaytesSMC.SMCBuffer — Type

struct SMCBuffer{M<:NamedTuple, I<:Integer, L<:Real, T<:AbstractFloat, F<:AbstractFloat, P, J}

Buffer struct to inplace store intermediate values.

Fields

cumweights::Vector{T} where T<:AbstractFloat

Buffer for cumulative particle weights at each iteration, such that they can be used again next iteration after resampling step.

parameter::BaytesCore.ModelParameterBuffer{M, L, T, I} where {M<:NamedTuple, I<:Integer, L<:Real, T<:AbstractFloat}

Buffer for model parameter and reweighting statistics.

correlation::BaytesSMC.CorrelationBuffer

Buffer to compute correlation between original and jittered parameter.

predictions::Vector

Predictions

jitterdiagnostics::Vector

Diagnostics of jitter steps. If not jittered this iteration (i.e., accepted == true), contains jitterdiagnostics from previous step.

BaytesSMC.SMCConstructor — Type

Callable struct to make initializing SMC sampler easier in sampling library.

Examples

BaytesSMC.SMCDefault — Type

struct SMCDefault{F<:Function, T<:BaytesCore.ResamplingMethod, B<:UpdateBool, C<:UpdateBool, U<:UpdateBool}

Default arguments for SMC constructor.

Fields

Ntuning::Int64

Number of tuning steps used when constructing sampler.

resamplingmethod::BaytesCore.ResamplingMethod

Threshold for resampling chains.

resamplingthreshold::Float64

Threshold for resampling chains. Set to 1.0 if resampling should always be applied.

jitterfun::Function

Function that is applied to correlation vector of parameter to determine if jittering can be stopped.

jitteradaption::UpdateBool

Boolean if fixed number of jittersteps are applied or jittering is based on parameter correlation.

jitterdiagnostics::UpdateBool

Boolean if diagnostics in jittersteps should be recorded in SMCDiagnostics.

jitterthreshold::Float64

Stopping threshold against jitterfun(correlation) of jittered model parameter.

jittermin::Int64

Minimum number of jittering steps.

jittermax::Int64

Maximum number of jittering steps.

generated::UpdateBool

Boolean if generate(_rng, objective) for corresponding model is stored in PF Diagnostics.

BaytesSMC.SMCDiagnostics — Type

struct SMCDiagnostics{P, J, T<:AbstractFloat, G} <: BaytesCore.AbstractDiagnostics

SMC Diagnostics container, including diagnostics of kernels used in jittering step.

Fields

base::BaytesCore.BaseDiagnostics{Vector{P}} where P

Diagnostics used for all Baytes kernels

ℓincrement::Float64

Weighted average of incremental log weights - can be used for marginal or log predictive likelihood computation.

ℓweights::Vector{Float64}

Cumulative log weights. If resampled, updated with jittered parameter - as weights will be used for tempering.

ℓweightsₜ::Vector{Float64}

Log weights at current iteration, used for resampling criterion. NOT updated with jittered theta parameter as ℓweightsₜ depends on t-1 and jittered theta parameter come in at next iteration.

ℓweightsₙ::Vector{Float64}

Normalized Log weights. They might be adjusted from previous iterations, so will differ from ℓweights. Will be set to log(1/N) if resampling applied.

jitterdiagnostics::Vector

Diagnostics of jitter steps. If not resampled this iteration (i.e., resampled == false), contains jitterdiagnostics from previous step.

jittersteps::Int64

Number of jittering steps

ρ::Vector{T} where T<:AbstractFloat

Correlation from rejuvented continuous parameter

ESS::Float64

ESS and accepted steps for SMC kernel

resampled::Bool

Boolean if step has been resampled.

generated::Any

Generated quantities specified for objective

BaytesSMC.SMCParticles — Type

struct SMCParticles{B<:BaytesCore.AbstractAlgorithm, A<:ModelWrappers.ModelWrapper, D<:SMCBuffer}

SMC kernel containing particles, dynamics and buffers for sampling.

Fields

model::Vector{A} where A<:ModelWrappers.ModelWrapper

SMC particles used for sampling.

kernel::Vector{B} where B<:BaytesCore.AbstractAlgorithm

Individual kernels associated to an SMC particle.

weights::BaytesCore.ParameterWeights

Weights struct for all SMC particles.

buffer::SMCBuffer

Buffer values for allocation free evaluations.

BaytesSMC.SMCParticles — Method

SMCParticles(_rng, JitterKernel, objective, info, tune)

Initialize SMC particles.

Examples

BaytesSMC.SMCParticles — Method

SMCParticles(_rng, kernel, objective, info, tune)

Initialize SMC² sampler.

Examples

BaytesSMC.SMCTune — Type

struct SMCTune{T<:ModelWrappers.Tagged, F<:Function, B<:UpdateBool, C<:UpdateBool, R<:BaytesCore.ResamplingMethod, U<:UpdateBool, S<:UpdateBool}

SMC tuning container.

Fields

tagged::ModelWrappers.Tagged

Tagged Model parameter.

iter::BaytesCore.Iterator

Iterator that contains current data iteration ~ starts with Ndata.

chains::BaytesCore.ChainsTune

ParticleTune struct to possibly adaptively change number of chains.

resample::BaytesCore.ResampleTune

Resampling Method of parameter particle indices.

jitter::BaytesCore.JitterTune

Function that is applied to correlation of parameter to determine if jittering can be stopped.

jitterfun::Function

Determines number of jittering steps in jitter kernel.

jitterdiagnostics::UpdateBool

Boolean if diagnostics in jittersteps should be recorded in SMCDiagnostics.

capture::UpdateBool

UpdateBool if parameter can be captured after initial jitter step.

Ntuning::Int64

Number of warmup Tuning steps for kernels.

batchsize::Int64

Batchsize for parallel processing of jittering steps.

generated::UpdateBool

Boolean if generated quantities should be generated while sampling

BaytesCore.infer — Method

infer(_rng, diagnostics, smc, model, data)

Infer MCMC diagnostics type.

Examples

BaytesCore.jitter! — Method

jitter!(_rng, particles, tune, data, temperature)

Jitter θ particles with given kernels. This is performed in 2 stages:

kernel is updated with new data and shuffled particles, then one proposal step is performed.

If more than one jitterstep has to be performed, previous results might be captured depending on the kernel.

Examples

BaytesCore.propagate! — Method

Propagate data forward over time. If (latent) data has to be extended, need to overload this function for specific smc kernel.

Examples

BaytesCore.propagate! — Method

propagate!(_rng, particles, tune, data, temperature)

Propagate data forward over time.

Examples

BaytesCore.propose! — Method

propose!(_rng, smc, model, data)
propose!(_rng, smc, model, data, temperature)
propose!(_rng, smc, model, data, temperature, update)

Propose new parameter with smc sampler. If update=true, objective function will be updated with input model and data.

Examples

BaytesCore.resample! — Method

Resample particle ancestors, shuffle current particles and rejuvenate θ.

Examples

BaytesCore.results — Method

results(diagnosticsᵛ, smc, Ndigits, quantiles)

Print diagnostics result of SMC sampler.

Examples

BaytesCore.weight! — Method

Compute new cumulative and incremental particle weights.

Examples

BaytesSMC.SMCreweight — Method

SMCreweight(_rng, objective, algorithm, cumweightsₜ₋₁)

Computes particle weights after jiterring step. Defaults to SMCweight function.

Examples

BaytesSMC.SMCweight — Method

SMCweight(_rng, objective, algorithm, cumweightsₜ₋₁)

Compute cumulative and incremental weight of objective at time/iteration t, given weight at t-1. Incremental weight will be used as particle weight for resampling. Cumulative weight will be used to adapt temperature.

If temperature is constant, this function can be overloaded with your ModelName if incremental weight can be computed independent of previous weight, which speeds up computation. cumweightsₜ is not needed in this case.

Examples

BaytesSMC.compute_batchsize — Function

compute_batchsize(Nchains)
compute_batchsize(Nchains, Nthreads)

Compute maximum number for batchsize in resampling step of SMC.

Examples

BaytesSMC.compute_ρ! — Method

compute_ρ!(ρ, pairs, valᵤ, algorithmᵛ)
compute_ρ!(ρ, pairs, valᵤ, algorithmᵛ, ϵ)

Compute inplace the correlation between old parameter vectors valᵤ and new parameter vectors from 'algorithmᵛ'.

Examples

BaytesSMC.infer_generated — Method

infer_generated(_rng, smc, model, data)

Infer type of generated quantities of PF sampler.

Examples

BaytesSMC.jitterkernel — Method

jitterkernel(particles, iter)

Return kernel that is used for jittering.

Examples

BaytesSMC.predict! — Method

Predict new data for each smc particle.

Examples

BaytesSMC.reweight! — Method

Compute new cumulative particle weights, accounting for jittering steps. This is only needed for next iteration's weight calculation, and will not adjust current incremental and normalized weight.

Examples